refactor(concrete-csprng): fixes visibility issue
+ Removes private types in the public api
+ Cleans the counter module
