+++ concrete-csprng/Cargo.toml
@@ -16,10 +16,10 @@ aes-soft = "0.6.4"
 
 [dev-dependencies]
 rand = "0.8.3"
+criterion = "0.3"
 
 [features]
 slow = []
-unchecked = []
 
 [[bin]]
 name = "generate_random"
